Add code for tutorial 02
