--- .02_multicore_rust/raspi3_boot/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "raspi3_boot"
-version = "0.1.0"
-authors = ["Andre Richter <andre.o.richter@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-panic-abort = "0.3.1"
-r0 = "0.2.2"
