+++ 02_runtime_init/src/bsp/rpi3.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+//
+// Copyright (c) 2018-2019 Andre Richter <andre.o.richter@gmail.com>
+
+//! Board Support Package for the Raspberry Pi 3.
+
+mod panic_wait;
+
+global_asm!(include_str!("rpi3/start.S"));
