+++ 02_runtime_init/src/bsp/rpi3/panic_wait.rs
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: MIT
+//
+// Copyright (c) 2018-2019 Andre Richter <andre.o.richter@gmail.com>
+
+//! A panic handler that infinitely waits.
+
+use core::panic::PanicInfo;
+
+#[panic_handler]
+fn panic(_info: &PanicInfo) -> ! {
+    unsafe {
+        loop {
+            asm!("wfe" :::: "volatile")
+        }
+    }
+}
